requirements.txt
absl-py==0.12.0
asgiref==3.3.4
asyncio==3.4.3
attrs==21.2.0
bidict==0.21.2
cmake==3.18.4.post1
dj-database-url==0.5.0
Django==3.2.2
django-colorfield==0.4.1
django-heroku==0.3.1
djangorestframework==3.12.4
dukpy==0.2.3
gunicorn==20.1.0
imutils==0.5.4
iteration-utilities==0.11.0
javascripthon==0.11
jsbuilder==0.1.0
macropy3==1.1.0b2
mediapipe==0.8.4.2
numpy==1.20.2
opencv-contrib-python==4.5.2.52
opencv-python==4.5.1.48
Pillow==8.2.0
python-engineio==4.1.0
python-socketio==5.2.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
tornado==6.1
typing-extensions==3.10.0.0
websocket-client==0.59.0
websockets==9.0.1
